separate screens, widgets, and services
